backend/requirement.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pgvector>=0.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
bcrypt==4.3.0